- pile of TODO updated - GameDatabase.java and EntityFactory.java more clearly refined - ClientGamePlay updated to reflect changes above

+++ RedShift/src/ents/BasicShot.java
@@ -76,7 +76,7 @@ public class BasicShot extends BaseEnt {
 		this.collider = collider;
 	}
 
-	public double getSpeed() {
+	public float getSpeed() {
 		return speed;
 	}
